requirements.txt
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cssselect==1.2.0
cssutils==2.6.0
Flask==2.2.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
premailer==3.10.0
PyJWT==2.6.0
python-dotenv
requests==2.28.2
SQLAlchemy==2.0.5.post1
typing-extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
WTForms==3.0.1
yagmail==0.15.293
zipp==3.15.0
